+++ rllab/envs/mujoco/hill/half_cheetah_hill_env.py
@@ -13,4 +13,4 @@ class HalfCheetahHillEnv(HillEnv):
     @overrides
     def _mod_hfield(self, hfield):
         # clear a flat patch for the robot to start off from
-        return terrain.clear_patch(hfield, gym.spaces.Box(np.array([-3.0, -1.5]), np.array([0.0, -0.5])))
+        return terrain.clear_patch(hfield, gym.spaces.Box(np.array([-3.0, -1.5]), np.array([0.0, -0.5])), dtype=np.float32)
