+++ rllab/envs/mujoco/hill/hopper_hill_env.py
@@ -13,4 +13,4 @@ class HopperHillEnv(HillEnv):
     @overrides
     def _mod_hfield(self, hfield):
         # clear a flat patch for the robot to start off from
-        return terrain.clear_patch(hfield, gym.spaces.Box(np.array([-1.0, -1.0]), np.array([-0.5, -0.5])))
+        return terrain.clear_patch(hfield, gym.spaces.Box(np.array([-1.0, -1.0]), np.array([-0.5, -0.5])), dtype=np.float32)
