+++ tests/algos/test_trpo.py
@@ -13,11 +13,11 @@ from rllab.policies import GaussianMLPPolicy
 class DummyEnv(gym.Env):
     @property
     def observation_space(self):
-        return gym.spaces.Box(low=-np.inf, high=np.inf, shape=(1, ))
+        return gym.spaces.Box(low=-np.inf, high=np.inf, shape=(1, ), dtype=np.float32)
 
     @property
     def action_space(self):
-        return gym.spaces.Box(low=-5.0, high=5.0, shape=(1, ))
+        return gym.spaces.Box(low=-5.0, high=5.0, shape=(1, ), dtype=np.float32)
 
     def reset(self):
         return np.zeros(1)
